Created File Reader for '*.dt1' type files.
I have provided the code in organized folders. The Documents folder contains information on the file format and example files.
The data layer contains data abstractions needed to parse the file and created subsequent data later. The code could use a 
little cleanup, but for now it works as intended!

This successful completes a milestone of reading in a '*.dt1' file and extracting the correct data.